requirements.txt
# pip3 install --user -r requirements.txt

dnspython==1.16.0
anytree==2.6.0
pytz==2019.1
asq==1.3
parse==1.12.0
netlib==0.17
IPy==1.0